requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.3
Flask-moment==1.0.6
mysql-connector-python==8.0.33
Flask-Excel==0.0.7
pyexcel-xlsx==0.6.0
Werkzeug==3.0.6
bcrypt==4.2.0
Flask-Caching==2.3.0
captcha==0.6.0
psutil==6.1.0
APScheduler==3.10.4
croniter==5.0.1
flasgger==0.9.7.1
Flask-CORS==3.0.10